+++ src/test/java/org/folio/rest/impl/TransactionTest.java
@@ -46,18 +46,17 @@ import io.restassured.http.Header;
 import io.vertx.core.json.JsonObject;
 
 class TransactionTest extends TestBase {
-  private static final String TRANSACTION_ENDPOINT = TRANSACTION.getEndpoint();
+  protected static final String TRANSACTION_ENDPOINT = TRANSACTION.getEndpoint();
   private static final String TRANSACTION_TEST_TENANT = "transaction_test_tenant";
-  private static final Header TRANSACTION_TENANT_HEADER = new Header(OKAPI_HEADER_TENANT, TRANSACTION_TEST_TENANT);
+  protected static final Header TRANSACTION_TENANT_HEADER = new Header(OKAPI_HEADER_TENANT, TRANSACTION_TEST_TENANT);
 
   private static final String FY_FUND_QUERY = "?query=fiscalYearId==%s AND fundId==%s";
   private static final String LEDGER_QUERY = "?query=fund.id==%s";
   private static final String LEDGER_FY_QUERY = "?query=ledgerId==%s AND fiscalYearId==%s";
   public static final String ALLOCATION_SAMPLE = "data/transactions/allocation_AFRICAHIST-FY20_ANZHIST-FY20.json";
   public static final String ENCUMBRANCE_SAMPLE = "data/transactions/encumbrance_AFRICAHIST_306857_1.json";
-  private static final String CREDIT_SAMPLE = "data/transactions-PaymentCredit/credit_CANHIST_30121.json";
-  private static final String PAYMENT_SAMPLE = "data/transactions-PaymentCredit/payment_ENDOW-SUBN_30121.json";
-  private static String BUDGETS_QUERY = BUDGET.getEndpoint() + FY_FUND_QUERY;
+
+  protected static String BUDGETS_QUERY = BUDGET.getEndpoint() + FY_FUND_QUERY;
   private static String LEDGERS_QUERY = LEDGER.getEndpoint() + LEDGER_QUERY;
   private static String LEDGER_FYS_ENDPOINT = getEndpoint(FinanceStorageLedgerFiscalYears.class) + LEDGER_FY_QUERY;
   private static final String BUDGETS = "budgets";
@@ -497,7 +496,7 @@ class TransactionTest extends TestBase {
 
   }
 
-  private void createOrderSummary(String orderId, int encumbranceNumber) throws MalformedURLException {
+  protected void createOrderSummary(String orderId, int encumbranceNumber) throws MalformedURLException {
     OrderTransactionSummary summary = new OrderTransactionSummary().withId(orderId).withNumTransactions(encumbranceNumber);
     postData(ORDER_TRANSACTION_SUMMARIES_ENDPOINT, JsonObject.mapFrom(summary)
       .encodePrettily(), TRANSACTION_TENANT_HEADER);
@@ -555,8 +554,6 @@ class TransactionTest extends TestBase {
   @Test
   void testCreateEncumbrancesDuplicateInTransactionTable() throws MalformedURLException {
 
-
-
     String orderId = UUID.randomUUID().toString();
     createOrderSummary(orderId,  2);
 
@@ -606,16 +603,10 @@ class TransactionTest extends TestBase {
     postData(TRANSACTION_ENDPOINT, transactionSample2, TRANSACTION_TENANT_HEADER)
       .then()
       .statusCode(201);
-
-
-
   }
 
   @Test
   void testUpdateEncumbranceAllOrNothing() throws MalformedURLException {
-
-
-
     String orderId = UUID.randomUUID().toString();
     createOrderSummary(orderId, 2);
 
@@ -696,15 +687,10 @@ class TransactionTest extends TestBase {
     assertEquals(expectedBudgetsAvailable , fromBudgetAfterUpdate.getAvailable());
     assertEquals(expectedBudgetsUnavailable, fromBudgetAfterUpdate.getUnavailable());
     assertEquals(expectedAwaitingPayment, fromBudgetAfterUpdate.getAwaitingPayment());
-
-
-
   }
 
   @Test
   void testUpdateAlreadyReleasedEncumbranceBudgetNotUpdated() throws MalformedURLException {
-
-
 
     String orderId = UUID.randomUUID().toString();
       createOrderSummary(orderId, 1);
@@ -748,17 +734,11 @@ class TransactionTest extends TestBase {
     assertEquals(fromBudgetBefore.getAvailable() , fromBudgetAfterUpdate.getAvailable());
     assertEquals(fromBudgetBefore.getUnavailable(), fromBudgetAfterUpdate.getUnavailable());
     assertEquals(fromBudgetBefore.getAwaitingPayment(), fromBudgetAfterUpdate.getAwaitingPayment());
-
-
-
   }
 
 
   @Test
   void testUpdateEncumbranceNotFound() throws MalformedURLException {
-
-
-
     String orderId = UUID.randomUUID().toString();
     createOrderSummary(orderId, 2);
 
@@ -769,222 +749,8 @@ class TransactionTest extends TestBase {
 
     // Try to update non-existent transaction
     putData(TRANSACTION.getEndpointWithId(), UUID.randomUUID().toString(), JsonObject.mapFrom(encumbrance).encodePrettily(), TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-
   }
 
-  @Test
-  void testCreatePaymentsCreditsAllOrNothing() throws MalformedURLException {
-
-
-    String invoiceId = UUID.randomUUID().toString();
-    String orderId = UUID.randomUUID().toString();
-    createOrderSummary(orderId, 2);
-    createInvoiceSummary(invoiceId, 2);
-    JsonObject jsonTx = new JsonObject(getFile(ENCUMBRANCE_SAMPLE));
-    jsonTx.remove("id");
-
-    Transaction paymentEncumbranceBefore = jsonTx.mapTo(Transaction.class);
-    paymentEncumbranceBefore.getEncumbrance().setSourcePurchaseOrderId(orderId);
-
-    Transaction creditEncumbranceBefore = jsonTx.mapTo(Transaction.class);
-    creditEncumbranceBefore.getEncumbrance().setSourcePurchaseOrderId(orderId);
-    creditEncumbranceBefore.getEncumbrance().setSourcePoLineId(UUID.randomUUID().toString());
-    String fY = paymentEncumbranceBefore.getFiscalYearId();
-    String fromFundId = paymentEncumbranceBefore.getFromFundId();
-
-
-    // prepare budget queries
-    String budgetEndpointWithQueryParams = String.format(BUDGETS_QUERY, fY, fromFundId);
-
-
-    // create 1st Encumbrance, expected number is 2
-    String paymentEncumbranceId = postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(paymentEncumbranceBefore).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    // create 2nd Encumbrance
-    String creditEncumbranceId = postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(creditEncumbranceBefore).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-      .statusCode(201)
-      .extract()
-      .as(Transaction.class).getId();
-
-    Budget budgetBefore = getBudgetAndValidate(budgetEndpointWithQueryParams);
-
-    JsonObject paymentjsonTx = new JsonObject(getFile(PAYMENT_SAMPLE));
-    paymentjsonTx.remove("id");
-
-    Transaction payment = paymentjsonTx.mapTo(Transaction.class);
-    payment.setSourceInvoiceId(invoiceId);
-    payment.setFiscalYearId(fY);
-    payment.setFromFundId(fromFundId);
-    payment.setPaymentEncumbranceId(paymentEncumbranceId);
-
-    String paymentId = postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(payment).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-    .statusCode(201)
-    .extract()
-    .as(Transaction.class).getId();
-
-    // payment does not appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), paymentId, TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-    JsonObject creditjsonTx = new JsonObject(getFile(CREDIT_SAMPLE));
-    creditjsonTx.remove("id");
-
-    Transaction credit = creditjsonTx.mapTo(Transaction.class);
-    credit.setSourceInvoiceId(invoiceId);
-    credit.setFiscalYearId(fY);
-    credit.setToFundId(fromFundId);
-    credit.setPaymentEncumbranceId(creditEncumbranceId);
-
-
-    String creditId = postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(credit).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-    .statusCode(201)
-    .extract()
-    .as(Transaction.class).getId();
-
-    // 2 transactions(each for a payment and credit) appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), paymentId, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-    getDataById(TRANSACTION.getEndpointWithId(), creditId, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-
-    Transaction paymentEncumbranceAfter = getDataById(TRANSACTION.getEndpointWithId(), paymentEncumbranceId, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-    Transaction creditEncumbranceAfter = getDataById(TRANSACTION.getEndpointWithId(), creditEncumbranceId, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-
-    //Encumbrance Changes for payment
-    assertEquals(-payment.getAmount(), subtractValues(paymentEncumbranceAfter.getEncumbrance().getAmountAwaitingPayment(), paymentEncumbranceBefore.getEncumbrance().getAmountAwaitingPayment()));
-    assertEquals(payment.getAmount(), subtractValues(paymentEncumbranceAfter.getEncumbrance().getAmountExpended(), paymentEncumbranceBefore.getEncumbrance().getAmountExpended()));
-    assertEquals(-payment.getAmount(), subtractValues(paymentEncumbranceAfter.getAmount(), paymentEncumbranceBefore.getAmount()));
-
-  //Encumbrance Changes for payment
-    assertEquals(-credit.getAmount(), subtractValues(creditEncumbranceAfter.getEncumbrance().getAmountExpended(), creditEncumbranceBefore.getEncumbrance().getAmountExpended()));
-
-
-    Budget budgetAfter = getBudgetAndValidate(budgetEndpointWithQueryParams);
-
-    //awaiting payment must decreases by payment amount
-    assertEquals(-payment.getAmount(), subtractValues(budgetAfter.getAwaitingPayment(), budgetBefore.getAwaitingPayment()));
-    //encumbered must increase by credit amount
-    assertEquals(credit.getAmount(), subtractValues(budgetAfter.getEncumbered(), budgetBefore.getEncumbered()));
-    //expenditures must increase by payment amt and decrease by credit amount
-    assertEquals(subtractValues(payment.getAmount(),credit.getAmount()), subtractValues(budgetAfter.getExpenditures(), budgetBefore.getExpenditures()));
-
-
-
-  }
-
-  @Test
-  void testCreatePaymentsCreditsAllOrNothingWithNoEncumbrances() throws MalformedURLException {
-
-
-    String invoiceId = UUID.randomUUID().toString();
-    createInvoiceSummary(invoiceId, 2);
-
-    JsonObject paymentjsonTx = new JsonObject(getFile(PAYMENT_SAMPLE));
-    paymentjsonTx.remove("id");
-
-    Transaction payment = paymentjsonTx.mapTo(Transaction.class);
-    payment.setSourceInvoiceId(invoiceId);
-    payment.setPaymentEncumbranceId(null);
-
-    String fY = payment.getFiscalYearId();
-    String fromFundId = payment.getFromFundId();
-
-
-    // prepare budget queries
-    String paymentBudgetEndpointWithQueryParams = String.format(BUDGETS_QUERY, fY, fromFundId);
-    Budget paymentBudgetBefore = getBudgetAndValidate(paymentBudgetEndpointWithQueryParams);
-
-    String paymentId = postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(payment).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-    .statusCode(201)
-    .extract()
-    .as(Transaction.class).getId();
-
-    // payment does not appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), paymentId, TRANSACTION_TENANT_HEADER).then().statusCode(404);
-
-    JsonObject creditjsonTx = new JsonObject(getFile(CREDIT_SAMPLE));
-    creditjsonTx.remove("id");
-
-    Transaction credit = creditjsonTx.mapTo(Transaction.class);
-    credit.setSourceInvoiceId(invoiceId);
-    credit.setFiscalYearId(fY);
-    credit.setPaymentEncumbranceId(null);
-    credit.setAmount(30d);
-
-    String creditBudgetEndpointWithQueryParams = String.format(BUDGETS_QUERY, fY, credit.getToFundId());
-    Budget creditBudgetBefore = getBudgetAndValidate(creditBudgetEndpointWithQueryParams);
-
-
-    String creditId = postData(TRANSACTION_ENDPOINT, JsonObject.mapFrom(credit).encodePrettily(), TRANSACTION_TENANT_HEADER).then()
-    .statusCode(201)
-    .extract()
-    .as(Transaction.class).getId();
-
-    // 2 transactions appear in transaction table
-    getDataById(TRANSACTION.getEndpointWithId(), paymentId, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-    getDataById(TRANSACTION.getEndpointWithId(), creditId, TRANSACTION_TENANT_HEADER).then().statusCode(200).extract().as(Transaction.class);
-
-
-    Budget paymentBudgetAfter = getBudgetAndValidate(paymentBudgetEndpointWithQueryParams);
-    Budget creditBudgetAfter = getBudgetAndValidate(creditBudgetEndpointWithQueryParams);
-
-
-    //awaiting payment, encumberedmexpenditures must remain same
-    assertEquals(0d, subtractValues(paymentBudgetAfter.getAwaitingPayment(), paymentBudgetBefore.getAwaitingPayment()));
-    assertEquals(0d, subtractValues(paymentBudgetAfter.getEncumbered(), paymentBudgetBefore.getEncumbered()));
-    assertEquals(0d, subtractValues(paymentBudgetAfter.getExpenditures(), paymentBudgetBefore.getExpenditures()));
-
-    //payment changes, available must decrease and unavailable increases
-    assertEquals(-payment.getAmount(), subtractValues(paymentBudgetAfter.getAvailable(), paymentBudgetBefore.getAvailable()));
-    assertEquals(payment.getAmount(), subtractValues(paymentBudgetAfter.getUnavailable(), paymentBudgetBefore.getUnavailable()));
-
-
-  //credit changes, available must increase and unavailable decreases
-    assertEquals(credit.getAmount(), subtractValues(creditBudgetAfter.getAvailable(), creditBudgetBefore.getAvailable()));
-    assertEquals(-credit.getAmount(), subtractValues(creditBudgetAfter.getUnavailable(), creditBudgetBefore.getUnavailable()));
-
-
-
-  }
-
-  @Test
-  void testCreatePaymentWithoutSummary() throws MalformedURLException {
-
-
-    String invoiceId = UUID.randomUUID().toString();
-
-    JsonObject jsonTx = new JsonObject(getFile(PAYMENT_SAMPLE));
-    jsonTx.remove("id");
-    Transaction payment = jsonTx.mapTo(Transaction.class);
-
-    payment.setSourceInvoiceId(invoiceId);
-
-    String transactionSample = JsonObject.mapFrom(payment).encodePrettily();
-
-    postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(400);
-
-
-
-  }
-
-  @Test
-  void testCreatePaymentsWithoutSummary() throws MalformedURLException {
-
-    String invoiceId = UUID.randomUUID().toString();
-
-    JsonObject jsonTx = new JsonObject(getFile(PAYMENT_SAMPLE));
-    jsonTx.remove("id");
-    Transaction payment = jsonTx.mapTo(Transaction.class);
-
-    payment.setSourceInvoiceId(invoiceId);
-    String transactionSample = JsonObject.mapFrom(payment).encodePrettily();
-
-    postData(TRANSACTION_ENDPOINT, transactionSample, TRANSACTION_TENANT_HEADER).then()
-      .statusCode(400);
-
-  }
 
 
   private Ledger getLedgerAndValidate(String endpoint) throws MalformedURLException {
@@ -1003,7 +769,7 @@ class TransactionTest extends TestBase {
       .as(LedgerFYCollection.class).getLedgerFY().get(0);
   }
 
-  private Budget getBudgetAndValidate(String endpoint) throws MalformedURLException {
+  protected Budget getBudgetAndValidate(String endpoint) throws MalformedURLException {
     return getData(endpoint, TRANSACTION_TENANT_HEADER).then()
       .statusCode(200)
       .body(BUDGETS, hasSize(1))
@@ -1011,7 +777,7 @@ class TransactionTest extends TestBase {
       .as(BudgetCollection.class).getBudgets().get(0);
   }
 
-  private double subtractValues(double d1, double d2) {
+  protected double subtractValues(double d1, double d2) {
     return BigDecimal.valueOf(d1).subtract(BigDecimal.valueOf(d2)).doubleValue();
   }
 
@@ -1019,7 +785,7 @@ class TransactionTest extends TestBase {
     return BigDecimal.valueOf(d1).add(BigDecimal.valueOf(d2)).doubleValue();
   }
 
-  private void createInvoiceSummary(String invoiceId, int numPaymentsCredits) throws MalformedURLException {
+  protected void createInvoiceSummary(String invoiceId, int numPaymentsCredits) throws MalformedURLException {
     InvoiceTransactionSummary summary = new InvoiceTransactionSummary().withId(invoiceId).withNumPaymentsCredits(numPaymentsCredits).withNumEncumbrances(0);
     postData(INVOICE_TRANSACTION_SUMMARIES_ENDPOINT, JsonObject.mapFrom(summary)
       .encodePrettily(), TRANSACTION_TENANT_HEADER);
